Revert oc-text-input to input as root
The oc-text-input was wrapped into a div in order to display a message
below the input. This creates problems in many places where the message
is not even needed. So this component is now being reverted into a pure
input component but with the addition of having the classes for a
warning or error state on the input border wired through properties.