Extract sort handling from OcTable
The sorting depends on the structure of resources which do not exist in ODS but only in web-app-files.
Moreover, we need to sort before pagination and as pagination is not handled in OcTable either, we need to handle sorting externally as well.
This has the additional benefit that this prepares us better for retrieving server sorted and paginated resources.