use single-storey letter a
use single storey letter "a"

what does it do?
go to https://rsms.me/inter/lab/?feat-cv11=1 and tick "cv 11 (Single-storey a)" and watch how the small letter "a" changes.

What does it improve?
The rendering of a single storey letter a aligns with our other rounded letters and digits.